app/backend/requirements.txt
doclayout_yolo==0.0.2
fastapi==0.115.5
Pillow==11.0.0
torch==2.4.0
uvicorn==0.32.0
dill==0.3.9
python-multipart
pypdfium2==4.30.0